updated structure of backend files
