--- mermaid-test.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Mermaid Test</title>
-    <style>
-        body {
-            background: #111827;
-            color: white;
-            font-family: system-ui, -apple-system, sans-serif;
-            padding: 20px;
-        }
-        .mermaid {
-            background: #1f2937;
-            padding: 20px;
-            border-radius: 8px;
-            margin: 20px 0;
-        }
-    </style>
-</head>
-<body>
-    <h1>Mermaid Diagram Test</h1>
-    
-    <h2>Test 1: Simple Flowchart</h2>
-    <div class="mermaid">
-        graph TD
-        A[Start] --> B{Decision}
-        B -->|Yes| C[End]
-        B -->|No| D[Other End]
-    </div>
-
-    <h2>Test 2: Repository Structure (From Backend)</h2>
-    <div class="mermaid">
-        graph TD
-        group_md[MD Files]
-        file_1[README.md]
-        group_md --> file_1
-        click file_1 "README.md"
-        group_json[JSON Files]
-        file_2[package.json]
-        group_json --> file_2
-        click file_2 "package.json"
-        group_md --> group_json
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/mermaid@10/dist/mermaid.min.js"></script>
-    <script>
-        mermaid.initialize({ 
-            startOnLoad: true,
-            theme: 'dark',
-            themeVariables: {
-                primaryColor: '#3b82f6',
-                primaryTextColor: '#f3f4f6',
-                primaryBorderColor: '#6b7280',
-                lineColor: '#9ca3af',
-                secondaryColor: '#1f2937',
-                tertiaryColor: '#374151',
-                background: '#111827',
-                darkMode: true
-            }
-        });
-    </script>
-</body>
-</html>
