+++ setup.py
@@ -1,0 +1,44 @@
+#!/usr/bin/env python3
+"""
+Setup configuration for the AI Code Architecture Agent
+"""
+from setuptools import setup, find_packages
+
+with open("README.md", "r", encoding="utf-8") as fh:
+    long_description = fh.read()
+
+with open("requirements.txt", "r", encoding="utf-8") as fh:
+    requirements = [line.strip() for line in fh if line.strip() and not line.startswith("#")]
+
+setup(
+    name="ai-code-architecture-agent",
+    version="1.0.0",
+    author="Jayasai Srikar",
+    description="AI-powered code analysis and architecture tool",
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    packages=find_packages(),
+    classifiers=[
+        "Development Status :: 4 - Beta",
+        "Intended Audience :: Developers",
+        "License :: OSI Approved :: MIT License",
+        "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
+    ],
+    python_requires=">=3.8",
+    install_requires=requirements,
+    entry_points={
+        "console_scripts": [
+            "ai-agent=main:main",
+        ],
+    },
+    package_data={
+        "src": ["static/*", "templates/*"],
+    },
+    include_package_data=True,
+)
