+++ src/__init__.py
@@ -1,0 +1,6 @@
+"""
+Codebase Agent - AI-powered code analysis and architecture tool
+"""
+
+__version__ = "1.0.0"
+__author__ = "Jayasai Srikar"
