+++ src/ai/__init__.py
@@ -1,0 +1,3 @@
+"""
+AI clients and conversation management
+"""
