+++ src/ai/gemini_client.py
