+++ src/api/__init__.py
@@ -1,0 +1,3 @@
+"""
+API endpoints and web interface
+"""
