+++ src/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core functionality for repository handling and diagram generation
+"""
