+++ src/core/gitdiagram_core_prompts.py
