+++ src/data/__init__.py
@@ -1,0 +1,3 @@
+"""
+Data management and knowledge base
+"""
