+++ src/scripts/__init__.py
@@ -1,0 +1,3 @@
+"""
+Development and debugging scripts
+"""
