+++ src/scripts/check_knowledge_base.py
