+++ src/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+Utility functions and helpers
+"""
